refactor: segment adapters by runtime
