+++ src/actions/AGENTS.md
@@ -1,6 +1,6 @@
 # src/actions
 
-> Last updated: 2026-07-29
+> Last updated: 2026-08-19
 
 ## Overview
 
@@ -24,7 +24,7 @@ src/actions/
   - 서비스 레이어가 던진 `AppError`를 액션이 다시 throw해서 `boundary.ts`의 `actionError`가 캐치·번역(`{ success:false, error }`)하는 것도 이 규칙 위반 아님 — 액션 "자신의" 검증 실패가 아니라 더 아래 레이어에서 이미 난 예외를 리턴값으로 옮기는 것뿐이다. 상세 규칙은 `docs/architecture/error-handling.md` §채널 A 참고.
 - 클라이언트가 넘긴 값을 소유권 판단 없이 그대로 DB 조회/수정 조건으로 쓰지 않는다 — 리소스 참조(ID)와 변경 내용만 클라이언트에서 받고, 소유자/권한은 세션에서 다시 조회해 대조한다. zod 등 스키마 검증은 값의 "형태"만 보장할 뿐 소유권을 보장하지 않는다.
 - DB 레코드를 그대로 반환값으로 넘기지 않는다 — Server Action의 리턴값은 클라이언트로 직렬화되므로, UI가 실제로 쓰는 필드만 추려 반환한다.
-- `models/`, `db/`, `adapters/`를 직접 import하지 않는다. 인증·외부 SDK·보상 로직과 업무 흐름 조합은 `services/`의 유스케이스 함수가 담당한다.
+- `models/`, `db/`, `adapters/`를 직접 import하지 않는다. 특히 `adapters/server/`도 직접 import하지 않는다. 인증·외부 SDK·보상 로직과 업무 흐름 조합은 `services/`의 유스케이스 함수가 담당한다.
 - mutation 이후 관련 캐시를 갱신하지 않고 끝내지 않는다 — Server Action 안에서는 즉시 반영(read-your-own-writes)이 필요하면 `updateTag`를, 오리진이 Route Handler 등 Server Action 바깥이면 `revalidateTag`/`revalidatePath`를 쓴다(`updateTag`는 Server Action 밖에서 호출하면 에러가 던져진다).
 - Client Component에서 직접 호출할 Server Action을 컴포넌트 파일 안에 인라인으로 정의하지 않는다 — Client Component는 `"use server"`가 선언된 별도 파일의 export만 import해 호출할 수 있다(인라인 함수 레벨 `"use server"`는 Server Component 전용).
 - `useActionState`로 연결되는 액션의 인자 순서를 `(prevState, formData)` 밖으로 바꾸지 않는다 — 이 훅의 계약이 이 순서를 요구한다.
