+++ src/adapters/browser/daum/index.ts
@@ -1,1 +1,3 @@
+import "client-only";
+
 export * from "./useDaumPopup";
