+++ src/adapters/browser/deeplink/index.ts
