+++ src/adapters/browser/deeplink/open-app.test.ts
