+++ src/adapters/browser/deeplink/open-app.ts
