+++ src/adapters/browser/kakao/index.ts
@@ -1,1 +1,3 @@
+import "client-only";
+
 export * from "./useKakaoLoader";
