+++ src/adapters/server/cloudinary/sign.test.ts
