+++ src/adapters/server/cookies/delete.ts
