+++ src/services/auth.ts
@@ -2,9 +2,9 @@ import "server-only";
 import type { UserRole } from "@/core/domain";
 import { UserModel } from "@/models";
 import { dbConnect } from "@/db";
-import { getCookie, deleteCookie, setCookie } from "@/adapters/cookies";
-import { decrypt, encrypt } from "@/adapters/jose";
-import { comparePasswords } from "@/adapters/bcrypt";
+import { getCookie, deleteCookie, setCookie } from "@/adapters/server/cookies";
+import { decrypt, encrypt } from "@/adapters/server/jose";
+import { comparePasswords } from "@/adapters/server/bcrypt";
 import mongoose from "mongoose";
 import { AppError } from "@/core/domain";
 import type { AuthSession } from "@/core/schemas";
