+++ vitest.config.ts
@@ -24,17 +24,13 @@ export default defineConfig({
           environment: "node",
           include: [
             "src/actions/**/*.test.ts",
-            "src/adapters/**/*.test.ts",
+            "src/adapters/server/**/*.test.ts",
             "src/core/**/*.test.ts",
             "src/app/api/**/*.test.ts",
             "src/boundary.test.ts",
             "src/proxy.test.ts",
           ],
-          exclude: [
-            "src/**/*.integration.test.{ts,tsx}",
-            "src/adapters/deeplink/**/*.test.ts",
-            "src/adapters/portone/request-payment.test.ts",
-          ],
+          exclude: ["src/**/*.integration.test.{ts,tsx}"],
           setupFiles: ["./testing/support/setup/node-polyfill.ts"],
           maxWorkers: 2,
         },
@@ -47,8 +43,7 @@ export default defineConfig({
           include: [
             "src/ui/**/*.test.{ts,tsx}",
             "src/app/**/*.test.{ts,tsx}",
-            "src/adapters/deeplink/**/*.test.ts",
-            "src/adapters/portone/request-payment.test.ts",
+            "src/adapters/browser/**/*.test.ts",
           ],
           exclude: ["src/**/*.integration.test.{ts,tsx}", "src/app/api/**/*.test.{ts,tsx}"],
           setupFiles: ["./testing/support/setup/jsdom-polyfill.ts"],
